testing/web-platform/tests/XMLHttpRequest/withcredentials-wrong-state.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: setting withCredentials when not in UNSENT, OPENED state</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-withcredentials-attribute" data-tested-assertations="following::ol/li[1] following::ol/li[2]" />  
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test(document.title)
      var client = new XMLHttpRequest()
      test.step(function() {
        client.open("GET", "resources/delay.py?ms=1000")
        client.send()
        assert_throws("InvalidStateError", function() { client.withCredentials=true })
        client.onreadystatechange = function(){
          test.step(function(){ 
            assert_throws("InvalidStateError", function() { client.withCredentials=true })
            if (client.readyState === 4) { 
              test.done()
            }
          }, "should throw when readyState is "+client.readyState)
        }
      }, "should throw when send flag is set")

    </script>
  </body>
</html>
